requirements.txt
asgiref==3.10.0
black==25.9.0
boto3==1.38.23
botocore==1.38.46
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.20
django-cors-headers==4.7.0
django-filter==25.1
django-rest-swagger==2.2.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
et_xmlfile==2.0.0
idna==3.11
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.3
mypy_extensions==1.1.0
numpy==2.3.3
openapi-codec==1.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.5.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytokens==0.2.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
ruff==0.14.1
s3transfer==0.13.1
simplejson==3.20.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
